long int -> int. não muda nada a principio.

+++ src/TAD.c
@@ -16,7 +16,7 @@ NODO_LSE *inserirLSE(NODO_LSE *lista, int num)
     return lista;
 }
 
-NODO_LSE *consultarLSE(NODO_LSE *lista, int num, long int *cmps)
+NODO_LSE *consultarLSE(NODO_LSE *lista, int num, int *cmps)
 {
     NODO_LSE *nodo = lista;
 
@@ -51,7 +51,7 @@ NODO_LSE *destruirLSE(NODO_LSE *lista)
 /*===============================================================================
                         Início - Árvore de Pesquisa Binária
 =================================================================================*/
-NODO_ABP *inserirABP(NODO_ABP *arv, int num, long int *cmps)
+NODO_ABP *inserirABP(NODO_ABP *arv, int num, int *cmps)
 {
     *cmps += 1;
     if (arv == NULL)
@@ -76,7 +76,7 @@ NODO_ABP *inserirABP(NODO_ABP *arv, int num, long int *cmps)
     return arv;
 }
 
-NODO_ABP *consultarABP(NODO_ABP *arv, int num, long int *cmps)
+NODO_ABP *consultarABP(NODO_ABP *arv, int num, int *cmps)
 {
     while (arv != NULL)
     {
@@ -136,7 +136,7 @@ NODO_AVL *rotacao_esquerda(NODO_AVL *pt)
     return pt;
 }
 
-NODO_AVL *rotacao_dupla_direita(NODO_AVL *pt, long int *cmps)
+NODO_AVL *rotacao_dupla_direita(NODO_AVL *pt, int *cmps)
 {
     NODO_AVL *ptu, *ptv;
 
@@ -159,7 +159,7 @@ NODO_AVL *rotacao_dupla_direita(NODO_AVL *pt, long int *cmps)
     return pt;
 }
 
-NODO_AVL *rotacao_dupla_esquerda(NODO_AVL *pt, long int *cmps)
+NODO_AVL *rotacao_dupla_esquerda(NODO_AVL *pt, int *cmps)
 {
     NODO_AVL *ptu, *ptv;
 
@@ -182,7 +182,7 @@ NODO_AVL *rotacao_dupla_esquerda(NODO_AVL *pt, long int *cmps)
     return pt;
 }
 
-NODO_AVL *caso1(NODO_AVL *arv, int *ok, long int *cmps)
+NODO_AVL *caso1(NODO_AVL *arv, int *ok, int *cmps)
 {
     NODO_AVL *ptu;
 
@@ -199,7 +199,7 @@ NODO_AVL *caso1(NODO_AVL *arv, int *ok, long int *cmps)
     return arv;
 }
 
-NODO_AVL *caso2(NODO_AVL *arv, int *ok, long int *cmps)
+NODO_AVL *caso2(NODO_AVL *arv, int *ok, int *cmps)
 {
     NODO_AVL *ptu;
 
@@ -216,7 +216,7 @@ NODO_AVL *caso2(NODO_AVL *arv, int *ok, long int *cmps)
     return arv;
 }
 
-NODO_AVL *inserirAVL(NODO_AVL *arv, int x, int *ok, long int *cmps)
+NODO_AVL *inserirAVL(NODO_AVL *arv, int x, int *ok, int *cmps)
 {
     /* Insere nodo em uma árvore AVL, onde A representa a raiz da árvore,
   x, a numave a ser inserida e h a altura da árvore */
@@ -279,7 +279,7 @@ NODO_AVL *inserirAVL(NODO_AVL *arv, int x, int *ok, long int *cmps)
     return arv;
 }
 
-NODO_AVL *consultarAVL(NODO_AVL *arv, int num, long int *cmps)
+NODO_AVL *consultarAVL(NODO_AVL *arv, int num, int *cmps)
 {
     while (arv != NULL)
     {
