+++ src/TAD.h
@@ -19,7 +19,7 @@ struct nodoLSE
 };
 
 NODO_LSE *inserirLSE(NODO_LSE *, int);
-NODO_LSE *consultarLSE(NODO_LSE *, int, long int *);
+NODO_LSE *consultarLSE(NODO_LSE *, int, int *);
 NODO_LSE *destruirLSE(NODO_LSE *);
 /*===============================================================================
                         Fim - Lista Simplesmente Encadeada 
@@ -36,8 +36,8 @@ struct nodoABP
     NODO_ABP *dir;
 };
 
-NODO_ABP *inserirABP(NODO_ABP *, int, long int *);
-NODO_ABP *consultarABP(NODO_ABP *, int, long int *);
+NODO_ABP *inserirABP(NODO_ABP *, int, int *);
+NODO_ABP *consultarABP(NODO_ABP *, int, int *);
 NODO_ABP *destruirABP(NODO_ABP *);
 /*===============================================================================
                         Fim - Árvore de Pesquisa Binária 
@@ -57,12 +57,12 @@ struct nodoAVL
 
 NODO_AVL *rotacao_direita(NODO_AVL *);
 NODO_AVL *rotacao_esquerda(NODO_AVL *);
-NODO_AVL *rotacao_dupla_direita(NODO_AVL *, long int *);
-NODO_AVL *rotacao_dupla_esquerda(NODO_AVL *, long int *);
-NODO_AVL *caso1(NODO_AVL *, int *, long int *);
-NODO_AVL *caso2(NODO_AVL *, int *, long int *);
-NODO_AVL *inserirAVL(NODO_AVL *, int, int *, long int *);
-NODO_AVL *consultarAVL(NODO_AVL *, int, long int *);
+NODO_AVL *rotacao_dupla_direita(NODO_AVL *, int *);
+NODO_AVL *rotacao_dupla_esquerda(NODO_AVL *, int *);
+NODO_AVL *caso1(NODO_AVL *, int *, int *);
+NODO_AVL *caso2(NODO_AVL *, int *, int *);
+NODO_AVL *inserirAVL(NODO_AVL *, int, int *, int *);
+NODO_AVL *consultarAVL(NODO_AVL *, int, int *);
 NODO_AVL *destruirAVL(NODO_AVL *);
 /*===============================================================================
                         Início - Árvore AVL
